using topPosts in final query to prevent error
given the latest sample code, the final query in this guide was generating an error. this could be fixed by removing the limit, or by requesting topPosts one level up on the viewer model. i went with the latter option as it also demonstrates the limit, rather than returning all posts without a limit.